backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlmodel==0.0.22
httpx==0.27.2
python-dotenv==1.0.1
google-generativeai>=0.3.0
python-dateutil>=2.8.0
backports.zoneinfo>=0.2.1; python_version<"3.9"
